backend/requirements.txt
flask==2.0.1
xrpl-py==1.5.0
gunicorn==20.1.0
flask-cors==3.0.10
